Fix division calculation in solution function

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.cpp
@@ -4,6 +4,6 @@
 using namespace std;
 
 int solution(int num1, int num2) {
-    int answer = ((double)num1/(double)num2) * 1000;
+    int answer = ((double)num1/num2) * 1000;
     return answer;
 }
